backend/requirements.txt
# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlmodel>=0.0.14

# Validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# LLM integration (Gemini 2.5 Pro)
google-generativeai>=0.3.0
